move tests to ScalarSuite
